requirements.txt
boto3==1.14.38
boto3-stubs==1.14.38.0
botocore==1.17.38
docutils==0.15.2
jmespath==0.10.0
msql==1.2.0
mypy==0.770
mypy-extensions==0.4.3
mysql-connector==2.2.9
psycopg2-binary==2.8.4
python-dateutil==2.8.1
s3transfer==0.3.3
six==1.15.0
typed-ast==1.4.1
typing-extensions==3.7.4.1
urllib3==1.25.10
